+++ narwhals/_pandas_like/expr.py
@@ -5,9 +5,9 @@ from typing import Any
 from typing import Callable
 from typing import Literal
 
+from narwhals._expression_parsing import reuse_series_implementation
+from narwhals._expression_parsing import reuse_series_namespace_implementation
 from narwhals._pandas_like.series import PandasSeries
-from narwhals._pandas_like.utils import reuse_series_implementation
-from narwhals._pandas_like.utils import reuse_series_namespace_implementation
 
 if TYPE_CHECKING:
     from typing_extensions import Self
@@ -50,6 +50,8 @@ class PandasExpr:
         from narwhals._pandas_like.namespace import PandasNamespace
 
         return PandasNamespace(self._implementation, self._backend_version)
+
+    def __narwhals_expr__(self) -> None: ...
 
     @classmethod
     def from_column_names(
