+++ narwhals/_pandas_like/group_by.py
@@ -9,9 +9,9 @@ from typing import Callable
 from typing import Iterable
 from typing import Iterator
 
+from narwhals._expression_parsing import parse_into_exprs
 from narwhals._pandas_like.utils import is_simple_aggregation
 from narwhals._pandas_like.utils import native_series_from_iterable
-from narwhals._pandas_like.utils import parse_into_exprs
 from narwhals.utils import remove_prefix
 
 if TYPE_CHECKING:
